% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream-package.R
\docType{package}
\name{coxstream-package}
\alias{coxstream-package}
\title{coxstream: Memory-Efficient Cox Proportional Hazards via Streaming Newton-Raphson}
\description{
Fits the Cox proportional hazards model using a single descending-order pass per Newton-Raphson iteration. Peak RAM is O(p^2) regardless of the number of rows, making it suitable for datasets that do not fit in memory. Produces identical coefficients to survival::coxph() with Efron tie correction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tommycarstensen/coxstream-r}
  \item Report bugs at \url{https://github.com/tommycarstensen/coxstream-r/issues}
}

}
\author{
\strong{Maintainer}: Tommy Carstensen \email{cran@tommycarstensen.com} (\href{https://orcid.org/0000-0002-3672-9931}{ORCID})

Authors:
\itemize{
  \item Tommy Carstensen \email{cran@tommycarstensen.com} (\href{https://orcid.org/0000-0002-3672-9931}{ORCID})
}

Other contributors:
\itemize{
  \item Apache Software Foundation (vendored Arrow C Data/Stream interface header (src/arrow_c_abi.h), Apache-2.0) [copyright holder, contributor]
}

}
\keyword{internal}
